EDA and regression analysis of H3.
EDA and regression analysis of H3. 
The document covers the description of each step and the specific explanation and analysis of the output. 
It also summarizes the results of the data study.
Tool: In case the teacher needs us to use R, I used Rstudio Mardown version, which will be easier to convert PDF.